move handlePackets for Linux out to separate package
